refactor: introduce `wallet/subnet/primary/common/utxotest`
