+++ wallet/subnet/primary/common/utxotest/utxotest.go
@@ -1,9 +1,7 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-//go:build test
-
-package common
+package utxotest
 
 import (
 	"context"
@@ -14,10 +12,11 @@ import (
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
+	"github.com/ava-labs/avalanchego/wallet/subnet/primary/common"
 )
 
 func NewDeterministicChainUTXOs(require *require.Assertions, utxoSets map[ids.ID][]*avax.UTXO) *DeterministicChainUTXOs {
-	globalUTXOs := NewUTXOs()
+	globalUTXOs := common.NewUTXOs()
 	for subnetID, utxos := range utxoSets {
 		for _, utxo := range utxos {
 			require.NoError(
@@ -26,12 +25,12 @@ func NewDeterministicChainUTXOs(require *require.Assertions, utxoSets map[ids.ID
 		}
 	}
 	return &DeterministicChainUTXOs{
-		ChainUTXOs: NewChainUTXOs(constants.PlatformChainID, globalUTXOs),
+		ChainUTXOs: common.NewChainUTXOs(constants.PlatformChainID, globalUTXOs),
 	}
 }
 
 type DeterministicChainUTXOs struct {
-	ChainUTXOs
+	common.ChainUTXOs
 }
 
 func (c *DeterministicChainUTXOs) UTXOs(ctx context.Context, sourceChainID ids.ID) ([]*avax.UTXO, error) {
